+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.0 (2023-11-13)
+
+- Rewrite harness and algorithms in plain C
+- Simplify test to a single function call
+- Set max bounce to 8 ticks
+- Test different noise frequencies
+- Add EvenOdd, FlipWait, and JustForFun algorithms
+- Change license to 0BSD
+
+
 ## 0.1.0 (2020-07-02)
 
 - Initial release.